[http] HttpWorker: fixes signed vs unsigned comparison in 413 content length check

+++ lib/http/HttpWorker.cpp
@@ -314,7 +314,7 @@ void HttpWorker::handleRequest(HttpRequest *r)
             r->finish();
             return;
         }
-        if (r->connection.contentLength() > server().maxRequestBodySize()) {
+        if (static_cast<size_t>(r->connection.contentLength()) > server().maxRequestBodySize()) {
             r->status = HttpStatus::RequestEntityTooLarge;
             r->finish();
             return;
